# api/swagger.json
{
  "openapi" : "3.0.0",
  "info" : {
    "version" : "0.4.0",
    "title" : "Trackless API",
    "contact" : {
      "name" : "wjtje",
      "url" : "https://www.wjtje.ga"
    },
    "license" : {
      "name" : "Apache 2.0 License",
      "url" : "https://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "paths" : {
...
    "/location" : {
      "get" : {
        "tags" : [ "location" ],
        "summary" : "Get all the locations on the server",
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/Location"
                  }
                }
              }
            }
          }
        },
        "security" : [ {
          "bearer_token" : [ ]
        } ]
      },
      "post" : {
        "tags" : [ "location" ],
        "summary" : "Create a new location",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "type" : "object",
                "properties" : {
                  "name" : {
                    "type" : "string"
                  },
                  "place" : {
                    "type" : "string"
                  },
                  "id" : {
                    "type" : "string"
                  }
                }
              }
            }
          }
        },
        "responses" : {
          "201" : {
            "description" : "Create a new location",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "properties" : {
                    "locationId" : {
                      "type" : "integer"
                    }
                  }
                }
              }
            }
          }
        },
        "security" : [ {
          "bearer_token" : [ ]
        } ]
      }
    },
    "/location/{locationId}" : {
      "get" : {
        "tags" : [ "location" ],
        "summary" : "Get infomation about a single location",
        "parameters" : [ {
          "in" : "path",
          "name" : "locationId",
          "description" : "ID of the location",
          "required" : true,
          "schema" : {
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/Location"
                  }
                }
              }
            }
          },
          "404" : {
            "description" : "Not found"
          }
        },
        "security" : [ {
          "bearer_token" : [ ]
        } ]
      },
      "delete" : {
        "tags" : [ "location" ],
        "summary" : "Remove a single location",
        "parameters" : [ {
          "in" : "path",
          "name" : "locationId",
          "description" : "ID of the location",
          "required" : true,
          "schema" : {
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation"
          },
          "404" : {
            "description" : "Not found"
          }
        },
        "security" : [ {
          "bearer_token" : [ ]
        } ]
      },
      "patch" : {
        "tags" : [ "location" ],
        "summary" : "Edit a location",
        "parameters" : [ {
          "in" : "path",
          "name" : "locationId",
          "description" : "ID of the location",
          "required" : true,
          "schema" : {
            "type" : "integer"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "type" : "object",
                "properties" : {
                  "name" : {
                    "type" : "string"
                  },
                  "place" : {
                    "type" : "string"
                  },
                  "id" : {
                    "type" : "string"
                  }
                }
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "successful operation"
          },
          "404" : {
            "description" : "Not found"
          }
        },
        "security" : [ {
          "bearer_token" : [ ]
        } ]
      }
    },
    "/location/user/{userId}/last" : {
      "get" : {
        "tags" : [ "location" ],
        "summary" : "Get the last used location",
        "parameters" : [ {
          "in" : "path",
          "name" : "userId",
          "description" : "ID of the user",
          "required" : true,
          "schema" : {
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/Location"
                  }
                }
              }
            }
          },
          "404" : {
            "description" : "Not found"
          }
        },
        "security" : [ {
          "bearer_token" : [ ]
        } ]
      }
    },
    "/location/user/{userId}/most" : {
      "get" : {
        "tags" : [ "location" ],
        "summary" : "Get the most used location",
        "parameters" : [ {
          "in" : "path",
          "name" : "userId",
          "description" : "ID of the user",
          "required" : true,
          "schema" : {
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/Location"
                  }
                }
              }
            }
          },
          "404" : {
            "description" : "Not found"
          }
        },
        "security" : [ {
          "bearer_token" : [ ]
        } ]
      }
    },
...
    "/user" : {
      "get" : {
        "tags" : [ "user" ],
        "summary" : "Get all the users from the system",
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/User"
                  }
                }
              }
            }
          }
        },
        "security" : [ {
          "bearer_token" : [ ]
        } ]
      },
      "post" : {
        "tags" : [ "user" ],
        "summary" : "Create a new user on the server",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "type" : "object",
                "properties" : {
                  "firstname" : {
                    "type" : "string"
                  },
                  "lastname" : {
                    "type" : "string"
                  },
                  "username" : {
                    "type" : "string"
                  },
                  "password" : {
                    "type" : "string"
                  },
                  "groupId" : {
                    "type" : "integer"
                  }
                }
              }
            }
          }
        },
        "responses" : {
          "201" : {
            "description" : "Created the new user",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "properties" : {
                    "userId" : {
                      "type" : "integer"
                    }
                  }
                }
              }
            }
          },
          "400" : {
            "description" : "Something is wrong with your input"
          }
        },
        "security" : [ {
          "bearer_token" : [ ]
        } ]
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "Location" : {
        "type" : "object",
        "properties" : {
          "locationId" : {
            "type" : "integer"
          },
          "name" : {
            "type" : "string"
          },
          "place" : {
            "type" : "string"
          },
          "id" : {
            "type" : "string"
          }
        }
      },
      "User" : {
        "type" : "object",
        "properties" : {
          "userId" : {
            "type" : "integer"
          },
          "firstname" : {
            "type" : "string"
          },
          "lastname" : {
            "type" : "string"
          },
          "username" : {
            "type" : "string"
          },
          "groupId" : {
            "type" : "integer"
          },
          "groupName" : {
            "type" : "string"
          }
        }
      },
...
    },
    "securitySchemes" : {
      "bearer_token" : {
        "type" : "apiKey",
        "name" : "api_key",
        "in" : "header"
      }
    }
  },
  "servers" : [ {
    "description" : "SwaggerHub API Auto Mocking",
    "url" : "https://virtserver.swaggerhub.com/wjtje/Trackless/0.4.0"
  } ]
}