# src/api/location/locationId.ts
import express from 'express';
import unusedRequestTypes from '../../scripts/RequestHandler/unusedRequestType';
import authHandler from '../../scripts/RequestHandler/authHandler';
import locationIdCheckHandler from '../../scripts/RequestHandler/idCheckHandler/locationIdCheckHandler';
import { DBcon } from '../..';
import { handleQuery } from '../../scripts/handle';
import { patchHandler, handlePatchQuery } from '../../scripts/RequestHandler/patchHandler';
import { mysqlTEXT } from '../../scripts/types';

const router = express.Router();

router.get(
  '/:locationId',
  authHandler('trackless.location.read'),
  locationIdCheckHandler(),
  (request, response, next) => {
    // Get the data from the server
    DBcon.query(
      "SELECT * FROM `TL_locations` WHERE `locationId`=?",
      [request.params.locationId],
      handleQuery(next, (result) => {
        response.status(200).json(result);
      })
    )
  }
)

router.delete(
  '/:locationId',
  authHandler('trackless.location.remove'),
  locationIdCheckHandler(),
  (request, response, next) => {
    // Delete from database
    DBcon.query(
      "DELETE FROM `TL_locations` WHERE `locationId`=?",
      [request.params.locationId],
      handleQuery(next, () => {
        response.status(200).json({
          message: 'done'
        })
      })
    );
  }
)

router.patch(
  '/:locationId',
  authHandler('trackless.location.edit'),
  locationIdCheckHandler(),
  patchHandler([
    {name: 'name', check: mysqlTEXT},
    {name: 'place', check: mysqlTEXT},
    {name: 'id', check: mysqlTEXT},
  ], (resolve, reject, key, request) => {
    // Update the key
    DBcon.query(
      "UPDATE `TL_locations` SET `" + key + "`=? WHERE `locationId`=?",
      [request.body[key], request.params.locationId],
      handlePatchQuery(reject, resolve)
    )
  })
)

router.use(unusedRequestTypes);

export default router;